* making properties public

+++ cyEngine/cyUtilities/source/cyMatrix2x2.cpp
@@ -3,7 +3,7 @@
 #include "cyMath.h"
 namespace CYLLENE_SDK {
   Matrix2x2::Matrix2x2(const float& value) {
-    memset(this, static_cast<int>(value), sizeof(Matrix2x2));
+    memset(this, value, sizeof(Matrix2x2));
   }
 
   Matrix2x2::Matrix2x2(const Matrix2x2& other)
